PR13283: Use a temporary string for comparisons
The reported issue was trying to directly compare the results of two
different kernel_string() calls, but they were always computed equal
even though they were known to be different.  Inspection of pass-3
showed that both calls led to an expr-stmt ending in ...->__retvalue.
This is just a pointer though, and it happens here that both calls will
point to the same string buffer, so strncmp rightly calls them equal.

To fix this, we must introduce temporary storage for the intermediate
strings.  Now we won't lose either value, and they can be properly
compared for equality.

* translate.cxx (c_tmpcounter::visit_comparison): Declare temporary
  storage for the left operand if it's not a literal.
  (c_unparser::visit_comparison): Use the left temporary.
* testsuite/systemtap.string/strcmp.*: Test string comparisons.