PR13274: Standardized Warning Processing (without elaborate.cxx)
Changed it so we always use session.print_warning() rather than
an explicit call to clog or cerr. (There are a few cases where
the session is not available, so they were left as is, but
marked up with an // XXX needs-session comment.)

Also changed print_warning() to check for suppress_warnings, so
we no longer have to explicitly check for it when calling
print_warning(). There are a few cases where I left in the
explicit check, specifically when there was more code than
just a simple call to print_warning() within it.

Also standardized the use of "WARNING:". All warnings should now
use "WARNING:" rather than "Warning:" or "warning:".