PR13489, the neverending kludge story.  Resolve RHEL5's old utrace conflicts.
RHEL5's old utrace uses utrace_native_view() instead of
task_user_regset_view(), so we shouldn't be so stuck on trying to find
it.  To add insult to injury, any attempt to use kallsyms* to look up
the latter results in an unloadable module since kallsyms* is not
exported in these old kernels.

* transport/transport.c (_stp_transport_init): Only look for
  task_user_regset_view for other than old rhel5 utrace.