Don't squash to pass 4 when using --remote
Our native_build check is a little too zealous, squashing runs to pass 4
when the local kernel doesn't match the command-line arguments.  When
we're running on a remote host, the local kernel is irrelevant.

This issue only cropped up if the user provided explicit -r/-a arguments
to match the remote -- otherwise the implicit searching was doing the
right thing with matching each kernel.

* session.cxx (systemtap_session::check_options): Remote targets don't
  care a whit whether the local kernel matches the build.