PR13546 Cont'd: Fixed failing test cases
Since we introduced a random seed to the hash function,
certain tests that printed results in an unordered
fashion could now fail, since the order is not necessarily
the same from run to run. These tests have been fixed and now
print in a certain order as to avoid those errors.