PR13609: Parallel Server Connections
Used pthreads to parallelize the server. A new thread is created to
handle each incomming connection. The maximum number of threads
created is limited by the new --max-threads=N option. If a value of
0 is passed, no new threads are created and each connection is
handled in the main thread, in serial. The default value is
equal to the number of processors on the host.