csclient: check directly for negative IPv4 pieces
As noted by clang, the code was looking for negative values of an
unsigned variable (which was read by strtoul).

  CXX    stap-csclient.o
../csclient.cxx:2638:44: error: comparison of unsigned expression < 0
      is always false [-Werror,-Wtautological-compare]
      if (errno != 0 || *estr != '\0' || p < 0 || p > 255)
                                         ~ ^ ~

This can never be negative, but strtoul will implicitly case negative-
looking strings to unsigned, which will make them fail p > 255 anyway.
But to be more direct, switch to [signed] long and use strtol.