PR13998: pretty-print bit fields numerically
When a struct member has bit attributes, we should always pretty-print
it as a number, even if the underlying type is a character.  Since these
attributes only show up in DWARF on the member, not the type die, we
need to treat it as a special case from the struct member iteration.

* tapsets.cxx (dwarf_pretty_print::recurse_bitfield): Print the incoming
  type numerically, regardless of whether it's a char.
  (dwarf_pretty_print::recurse_struct_members): If a member has a
  bit_offset attribute, use recurse_bitfield instead of normal recurse.
* testsuite/systemtap.printf/pretty-bits.*: Test how bitfields are
  pretty-printed, both for ints and chars underneath.