PR14146: adapted manpager.sh to produce passable tapset::* manpages.
The resulting manpages can act as 'very succinct' (read: utterly
bare-bones) tables of contents listing documented functions and probes,
and referring the user to probe::* and function::* manpages generated by
the 'robust' (read: hideously complicated and slow, but very reliable)
DocBook toolchain.

manpager.sh is currently not hooked up to the build sequence, pending
delivery of a fresh supply of fingers to cross in hope that everything
works as advertised.