PR13667: Hardcoded netfilter constants, cleaned up
Since certain older versions of the kernel do not define some
netfilter constants, we had to hardcode all the values
so that it would compile on those older kernels (i.e. RHEL 4 and 5).
This is OK, since we were only including the netfilter_* headers
to map the constants to their equivalent values.