PR13516 pre: Re-worked tmpdir creation/removal
In preparation for PR13516, we needed to rework the tmpdir
creation/removal, so as to be sure the tmpdir gets removed
when we get an exception. Creation now happens in the
systemtap_session ctor with a call to
systemtap_session::create_tmp_dir() and removal in the dtor
with a call to systemtap_session::remove_tmp_dir().

passes_0_4_again_with_server() now calls the new
systemtap_session::reset_tmp_dir(), rather than just removing
and relying on passes_0_4() to recreate it.