testsuite: Fix pretty-bits/char to work with relative paths
This is more fallout from the automake update (commit c66be968a4ea)
which led to relative $srcdir.  The pretty-bits and pretty-char tests
are passing $srcdir headers to @cast kernel types.  That gets compiled
with Kbuild, with the working directory is changed to the kernel root,
so we need to give full paths for those headers.

This adds a new tcl proc "fullpath" which inserts [pwd]/ on relative
paths, and updates those pretty tests to use it.