PR14244: Assert that the real UID has access to debugfs
Some systems are now defaulting debugfs to mode 0700.  If the user has
no access to debugfs, then they won't be able to communicate with the
systemtap module.  Thus, this should be a non-starter, before even
loading the module at all.

This was previously causing staprun to fail after loading the module,
when it checked R/W access to the module's .ctl file.  But since we also
restrict removing modules to those you can control, staprun would not
unload the new module either.